user/requirements/requirements.txt
boto3==1.11.0
botocore==1.14.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-cors-headers==3.1.1
django-enumfield==2.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
dropbox==9.4.0
google-api-python-client==1.7.11
google-auth==1.8.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
image==1.5.28
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
mysqlclient==1.4.4
Naked==0.1.31
oauth2client==4.1.3
oauthlib==3.1.0
onetimepass==1.0.1
openapi-codec==1.3.2
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycrypto==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.0
shellescape==3.8.1
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
